+++ example/component/accordion.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group'
+import { TransitionGroup, CSSTransition } from 'react-transition-group';
 
 class Accordion extends Component {
     static defaultProps = {
@@ -7,41 +7,40 @@ class Accordion extends Component {
     }
 
     constructor(props){
-        super(props)
+        super(props);
         this.state = {
             showContent: false,
             headerOpacity: 1
-        }
+        };
     }
 
     handleClick(e){
         this.setState({
             showContent: !this.state.showContent,
             headerOpacity: this.state.showContent ? 1 : 0.4
-        })
+        });
     }
 
     render() {
-        const { children, header, transitionName } = this.props
-        let content = this.state.showContent ? children : false
+        const { children, header, transitionName } = this.props;
+        let content = this.state.showContent ? children : <div></div>;
         return (
             <div>
                 <div onClick={this.handleClick.bind(this)}>
                     <div style={{
                         opacity: this.state.headerOpacity,
-                        transition : 'opacity .3s'
+                        transition: 'opacity .3s'
                     }}>{ header }</div>
                 </div>
-                <ReactCSSTransitionGroup
-                  transitionName={transitionName}
-                  transitionAppear={true}
-                  transitionAppearTimeout={500}
-                  transitionEnterTimeout={300}
-                  transitionLeaveTimeout={0}>
+                <CSSTransition
+                  classNames={transitionName}
+                  appear={this.state.showContent}
+                  in={this.state.showContent}
+                  timeout={{ enter: 300, exit: 0, appear: 500}}>
                     {content}
-                </ReactCSSTransitionGroup>
+                </CSSTransition>
             </div>
-        )
+        );
     }
 }