+++ example/pages/home/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Flex, FlexItem, Cells, Cell, CellBody, CellFooter} from '../../../src/index';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import Page from '../../component/page';
 import Accordion from '../../component/accordion';