+++ example/pages/msg/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 import { Button } from '../../../src/index';
 import Page from '../../component/page';
 
@@ -8,6 +8,6 @@ const MsgDemo = () => (
         <Button component={Link} to="/msg/success">Success Message</Button>
         <Button component={Link} to="/msg/fail">Warn Message</Button>
     </Page>
-)
+);
 
 export default MsgDemo;