+++ example/pages/page/index.js
@@ -3,7 +3,7 @@ import { Article, Page } from '../../../src/index';
 import srcArticle from '../article/pic_article.png';
 
 const PageDemo = (props) => (
-    <Page>
+    <Page transition={true} infiniteLoader={true} ptr={false}>
         <Article>
             <h1>Page Demo</h1>
             <section>