+++ src/components/form/preview_item.js
@@ -1,6 +1,7 @@
 //1.0.0 components
 
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import classNames from '../../utils/classnames';
 
 /**